refactor: main background fetch
