+++ src/components/background/MainBackground.tsx
@@ -3,63 +3,10 @@ import { ImageBackground, Dimensions, View, StyleSheet } from 'react-native';
 // import { Video, ResizeMode } from 'expo-av';
 import Video from 'react-native-video';
 import { useNoxSetting } from '@stores/useApp';
-import { fetchVideoPlayUrl } from '@utils/mediafetch/bilivideo';
 import { customReqHeader } from '@utils/BiliFetch';
-import { biliNFTVideoFetch } from '@utils/mediafetch/biliNFT';
-import { biliGarbHeadVideoFetch } from '@utils/mediafetch/biliGarb';
 import { logger } from '@utils/Logger';
 import { useIsLandscape } from '@hooks/useOrientation';
-import NoxCache from '@utils/Cache';
-
-enum RESOLVE_TYPE {
-  bvid = 'bvid',
-  video = 'video',
-  biliNFTVideo = 'biliNFTVideo',
-  biliGarbHeadVideo = 'biliGarbHeadVideo',
-  image = 'image',
-}
-
-export const resolveBackgroundImage = async (
-  backgroundImage: string | NoxTheme.backgroundImage
-) => {
-  if (typeof backgroundImage === 'string') {
-    return { type: RESOLVE_TYPE.image, identifier: backgroundImage };
-  }
-  switch (backgroundImage.type) {
-    case RESOLVE_TYPE.bvid:
-      return {
-        type: RESOLVE_TYPE.video,
-        identifier: await NoxCache.noxMediaCache?.loadCacheFunction(
-          `${RESOLVE_TYPE.bvid}-${backgroundImage.identifier}`,
-          () => fetchVideoPlayUrl(backgroundImage.identifier)
-        ),
-      };
-    case RESOLVE_TYPE.biliNFTVideo: {
-      const [act_id, index] = JSON.parse(backgroundImage.identifier);
-      return {
-        type: RESOLVE_TYPE.video,
-        identifier: await NoxCache.noxMediaCache?.loadCacheFunction(
-          `${RESOLVE_TYPE.biliNFTVideo}-${backgroundImage.identifier}`,
-          () => biliNFTVideoFetch({ act_id, index })
-        ),
-      };
-    }
-    case RESOLVE_TYPE.biliGarbHeadVideo:
-      return {
-        type: RESOLVE_TYPE.video,
-        identifier: await NoxCache.noxMediaCache?.loadCacheFunction(
-          `${RESOLVE_TYPE.biliGarbHeadVideo}-${backgroundImage.identifier}`,
-          () =>
-            biliGarbHeadVideoFetch({
-              act_id: backgroundImage.identifier,
-            })
-        ),
-      };
-
-    default:
-      return backgroundImage;
-  }
-};
+import { RESOLVE_TYPE } from '@utils/mediafetch/mainbackgroundfetch';
 
 const MainBackground = ({ children }: { children: React.JSX.Element }) => {
   const playerStyle = useNoxSetting(state => state.playerStyle);
